set up side bar to have sections
