+++ pages/Library_Sample_Level_Metadata.py
@@ -1,25 +1,25 @@
 import streamlit as st
 from src.format_page import render_header
 from src.field_matcher import load_data
-from src.transformer import transform_specimen_info
+from src.transformer import transform_library_sample_info
 from src.utils import load_schema
 
-session_name = "specimen_info"
-title = "specimen level metadata"
+session_name = "library_sample_info"
+title = "library sample level metadata"
 
 
-class SpecimenMetadataPage:
+class LibrarySampleMetadataPageMetadataPage:
     def __init__(
         self,
-        spec_required_fields,
-        spec_required_alternate_fields,
-        spec_optional_fields,
-        spec_optional_alternate_fields,
+        required_fields,
+        required_alternate_fields,
+        optional_fields,
+        optional_alternate_fields,
     ):
-        self.required_fields = spec_required_fields
-        self.required_alternate_fields = spec_required_alternate_fields
-        self.optional_fields = spec_optional_fields
-        self.optional_alternate_fields = spec_optional_alternate_fields
+        self.required_fields = required_fields
+        self.required_alternate_fields = required_alternate_fields
+        self.optional_fields = optional_fields
+        self.optional_alternate_fields = optional_alternate_fields
 
     def transform_and_save_data(
         self, df, mapped_fields, selected_optional_fields, selected_additional_fields
@@ -27,17 +27,17 @@ class SpecimenMetadataPage:
         if mapped_fields and selected_optional_fields != "Error":
             st.subheader("Transform Data")
             if st.button("Transform Data"):
-                transformed_df = transform_specimen_info(
+                transformed_df = transform_library_sample_info(
                     df.astype(object),
                     mapped_fields,
                     selected_optional_fields,
                     selected_additional_fields,
                 )
-                st.session_state[session_name] = transformed_df
+                st.session_state["library_sample_info"] = transformed_df
                 try:
-                    st.success("Specimen Information has been saved!")
+                    st.success("Library Sample Information has been saved!")
                 except Exception as e:
-                    st.error(f"Error saving Specimen Information: {e}")
+                    st.error(f"Error saving Library Sample Information: {e}")
 
     def display_panel_info(self, toggle_text):
         if session_name in st.session_state:
@@ -47,6 +47,7 @@ class SpecimenMetadataPage:
                 st.json(st.session_state[session_name])
 
     def run(self):
+        # File upload
         (
             df,
             mapped_fields,
@@ -59,27 +60,26 @@ class SpecimenMetadataPage:
             optional_alternate_fields,
         )
         # Transform and save data
-        if mapped_fields:
-            self.transform_and_save_data(
-                df, mapped_fields, selected_optional_fields, selected_additional_fields
-            )
+        self.transform_and_save_data(
+            df, mapped_fields, selected_optional_fields, selected_additional_fields
+        )
         # Display current panel information
         self.display_panel_info(f"Preview {title}")
 
 
-if __name__ == "__main__":
+if __name__ in ("__main__", "__page__"):
     render_header()
-    st.subheader("Specimen Level Metadata Converter", divider="gray")
+    st.subheader("Library Sample Level Metadata Converter", divider="gray")
     schema_fields = load_schema()
-    required_fields = schema_fields["specimen_level_metadata"]["required"]
-    required_alternate_fields = schema_fields["specimen_level_metadata"][
+    required_fields = schema_fields["library_sample_level_metadata"]["required"]
+    required_alternate_fields = schema_fields["library_sample_level_metadata"][
         "required_alternatives"
     ]
-    optional_fields = schema_fields["specimen_level_metadata"]["optional"]
-    optional_alternate_fields = schema_fields["specimen_level_metadata"][
+    optional_fields = schema_fields["library_sample_level_metadata"]["optional"]
+    optional_alternate_fields = schema_fields["library_sample_level_metadata"][
         "optional_alternatives"
     ]
-    app = SpecimenMetadataPage(
+    app = LibrarySampleMetadataPageMetadataPage(
         required_fields,
         required_alternate_fields,
         optional_fields,
